requirements.txt
launchpadlib
pyyaml
